requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
diskcache==5.6.3
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
hf-xet==1.1.2
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.42
llama_cpp_python==0.3.9
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==0.22.0
pikepdf==9.7.0
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.5.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
timm==1.0.15
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.52.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.35.0
unstructured-inference==1.0.2
unstructured.pytesseract==0.3.15
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.3
yarl==1.20.0
zstandard==0.23.0
flask-swagger-ui==4.11.1
